Basic Tagging has been implemented and the receipt class has been moulded to fit.
totalPrice is the highest price on the receipt.
dateOfPurchase is the first date on the receipt.
BussinessName is the largest TextBlock on the receipt.
OCR is a SparseArray of TextBlocks that we can iterate through when searching for product names.
image is the same, a the image taken.

New Constructor takes a SparseArray of TextBlocks from the OCR and an image and creates a receipt class.
This new constructor may mean that some of the members of the Receipt class do not get initialised. eg No date on the receipt.
You will need to handle null objects.

Minor lowered min fps to support more devices.